Added functionality to find "British Summer" champ
Functionality can find a "British Summer Championships" meet for years where the meet name had "British Summer" in it.
Support needs to be added so that it recognises "Aquatics GB" and "British Gas" nationals too, as well as the Trials held in Easter each year.
Intended purpose is so that we can easily scrape data from these elite domestic meets to compare stats on the performers at them through the age groups and gender such that analysis can be done, such as finding what strokes the best IM performers tend to be strong in and how this varies by age and gender, as well as how this varies for medalists each year as opposed to those who only made finals or those who only qualified. 
Aim is to plot these comparisons, perhaps using pie charts.