Threading for scraping best events of IM swimmers
Currently makes use of threading to scrape data from rankings such as 
- top ten swimmers in medley events
- their best 3 non-medley events
Parameters allow for customization as to how this data is scraped, such as:
- which years to scrape, a specific age group
- scrape a swimmers' best events in the season that they were in the top ten or just scrape their all time best events
- scrape a swimmer for multiple seasons if going for the approach of getting their best non-medley events per season rather than all time
Next steps are to add pie chart functionality and a parity parameter such that instead of finding top 3 events of a swimmer, it can find their best stroke by taking majority parity i.e. 2 strokes of the 3 events found. Can also do this for distance to find out if more sprinters are medley swimmers, more mid distance are medley, or distance, etc.