Medley scraper plots detailed pie charts for M&F
Plots for mens and womens 200 medley work as expected, with percentages on the slices, colour gradients matching with stroke and distance, and annotations on the charts.
Notable trend is that the most common non-medley event in top performing men is the 400 freestyle as opposed to the womens 200 freestyle. Also to note is that no women have 50 fly in their top 3 events whereas 1% of swims recorded for mens top 3 included the 50 fly.