More plots for medley swimmers; norm distributions
Added more flexibility with pie plotting for medley swimmers such that we can compare medley swimmers of the same generation, i.e people born in the same year can be compared from 12-13-14 etc.
Tried plots with different windows, i.e. taking teenage medley swimmers' best non-medley events for a 7 year timeframe, vs 10, vs 20 year, to see how trends change as the sample size is increased temporally. 
Added functionality for plotting normal distributions of events - need to improve it such that one can plot multiple distributions on the same plot to compare men and women, or two different events, or to be able to compare year on year to identify which years had better swimmers of a certain stroke/distance than others and for which age groups.