PCA of 200m IM
Not included in the coursework upload. The data shows the greatest varying components for 200m medley swimmers, with it clearly being butterfly and breaststroke, indicating that for a medley swimmer to be good, they should master one or both of these as these strokes cause the greatest variance, based on percentage of total swim time taken, across multiple elite athletes. Somewhat niche and not the most straightforward or usable data analysis, hence not being included in my coursework assignment.